+++ CHANGE_LOG.md
@@ -1,3 +1,24 @@
+## [0.72.1](https://github.com/oaknational/Oak-Web-Application/compare/v0.72.0...v0.72.1) (2022-08-22)
+
+
+### Bug Fixes
+
+* close menu on page navigation ([31cad11](https://github.com/oaknational/Oak-Web-Application/commit/31cad11750256243b51ef0a71dc58903cff76198))
+* put reference to menu context hooks in provider ([4620688](https://github.com/oaknational/Oak-Web-Application/commit/46206883fb8044717f19bc502e3662ed068ab5b7))
+
+# [0.72.0](https://github.com/oaknational/Oak-Web-Application/compare/v0.71.1...v0.72.0) (2022-08-22)
+
+
+### Bug Fixes
+
+* serialize policy date to ISO string, render w/ locale ([23fc363](https://github.com/oaknational/Oak-Web-Application/commit/23fc363483bd1169159d665bce9232484a4e9901))
+
+
+### Features
+
+* **cms:** add policyPageBySlug and policyPage methods to CMSClient ([6207cc6](https://github.com/oaknational/Oak-Web-Application/commit/6207cc6e1f071ddc95e8229e0ebe2b3f83ed3805))
+* **cms:** integrate policy pages with CMS ([8e63ea8](https://github.com/oaknational/Oak-Web-Application/commit/8e63ea83334e6a5b66d032828895b398a535e9c7))
+
 # [0.71.0](https://github.com/oaknational/Oak-Web-Application/compare/v0.70.1...v0.71.0) (2022-08-22)